Input bug here - Keyboard and mouse hold events not implemented yet * NEW: Raw Input - first workable version * DEL: DirectInput

+++ code/engine.vc2008/xrEngine/FDemoRecord.h
@@ -53,8 +53,8 @@ public:
 				CDemoRecord				(const char *name, float life_time=60*60*1000);
 	virtual		~CDemoRecord();
 
-	virtual void IR_OnKeyboardPress		(int dik);
-	virtual void IR_OnKeyboardHold		(int dik);
+	virtual void IR_OnKeyboardPress		(u8 dik);
+	virtual void IR_OnKeyboardHold		(u8 dik);
 	virtual void IR_OnMouseMove			(int dx, int dy);
 	virtual void IR_OnMouseHold			(int btn);
