+++ code/engine.vc2008/xrEngine/edit_actions.h
@@ -52,13 +52,13 @@ protected:
 class type_pair : public base
 {
 public:
-					type_pair		( u32 dik, char c, char c_shift, bool b_translate );
+					type_pair		( u8 dik, char c, char c_shift, bool b_translate );
 	virtual			~type_pair		();
-			void	init			( u32 dik, char c, char c_shift, bool b_translate );
+			void init ( u8 dik, char c, char c_shift, bool b_translate );
 	virtual	void	on_key_press	( line_edit_control* const control );
 
 private:
-	u32		m_dik;
+	u8		m_dik;
 	bool	m_translate;
 	char	m_char;
 	char	m_char_shift;
