+++ code/engine.vc2008/xrGame/Actor.h
@@ -415,9 +415,9 @@ public:
 	//////////////////////////////////////////////////////////////////////////
 public:
 	virtual void			IR_OnMouseMove			(int x, int y);
-	virtual void			IR_OnKeyboardPress		(int dik);
-	virtual void			IR_OnKeyboardRelease	(int dik);
-	virtual void			IR_OnKeyboardHold		(int dik);
+	virtual void			IR_OnKeyboardPress		(u8 dik);
+	virtual void			IR_OnKeyboardRelease	(u8 dik);
+	virtual void			IR_OnKeyboardHold		(u8 dik);
 	virtual void			IR_OnMouseWheel			(int direction);
 	virtual	float			GetLookFactor			();
