+++ code/engine.vc2008/xrGame/ai/monsters/snork/snork.h
@@ -34,7 +34,7 @@ public:
 
 private:
 #ifdef DEBUG	
-	virtual void	debug_on_key		(int key);
+	virtual void	debug_on_key		(u8 key);
 #endif
 
 public:
