+++ code/engine.vc2008/xrGame/ui/UIActorMenu.h
@@ -318,7 +318,7 @@ public:
 	virtual void				Update						();
 	virtual void				Show						(bool status);
 
-	virtual bool				OnKeyboardAction			(int dik, EUIMessages keyboard_action);
+	virtual bool				OnKeyboardAction			(u8 dik, EUIMessages keyboard_action);
 	virtual bool				OnMouseAction				(float x, float y, EUIMessages mouse_action);
 
 	void						CallMessageBoxYesNo			(LPCSTR text);
