+++ code/engine.vc2008/xrGame/ui/UIActorMenu_action.cpp
@@ -307,7 +307,7 @@ bool CUIActorMenu::OnMouseAction( float x, float y, EUIMessages mouse_action )
 	return true; // no click`s
 }
 
-bool CUIActorMenu::OnKeyboardAction(int dik, EUIMessages keyboard_action)
+bool CUIActorMenu::OnKeyboardAction(u8 dik, EUIMessages keyboard_action)
 {
 	InfoCurItem( NULL );
 	if ( is_binded(kDROP, dik) )
