+++ code/engine.vc2008/xrGame/ui/UICellItem.h
@@ -46,7 +46,7 @@ public:
 
 	
 
-	virtual		bool		OnKeyboardAction				(int dik, EUIMessages keyboard_action);
+	virtual		bool		OnKeyboardAction				(u8 dik, EUIMessages keyboard_action);
 	virtual		bool		OnMouseAction					(float x, float y, EUIMessages mouse_action);
 	virtual		void		Draw					();
 	virtual		void		Update					();
@@ -59,7 +59,7 @@ public:
 				bool		HasChild				(CUICellItem* item);
 	virtual		bool		EqualTo					(CUICellItem* itm);
 	IC const	Ivector2&	GetGridSize				()						{return m_grid_size;}; //size in grid
-	IC			void		SetAccelerator			(int dik)				{m_accelerator=dik;};
+	IC			void		SetAccelerator			(u8 dik)				{m_accelerator=dik;};
 	IC			int			GetAccelerator			()		const			{return m_accelerator;};
 
 	virtual		CUIDragItem* CreateDragItem			();
