+++ code/engine.vc2008/xrGame/ui/UIPdaWnd.cpp
@@ -307,7 +307,7 @@ void RearrangeTabButtons(CUITabControl* pTab)
 	pTab->SetWndPos( pos );
 }
 
-bool CUIPdaWnd::OnKeyboardAction(int dik, EUIMessages keyboard_action)
+bool CUIPdaWnd::OnKeyboardAction(u8 dik, EUIMessages keyboard_action)
 {
 	if (is_binded(kACTIVE_JOBS, dik))
 	{
