+++ code/engine.vc2008/xrGame/ui/uiscriptwnd_script.h
@@ -5,9 +5,9 @@ struct CWrapperBase : public T, public luabind::wrap_base {
 	typedef T inherited;
 	typedef CWrapperBase<T>	self_type;
 
-	virtual bool OnKeyboardAction(int dik, EUIMessages keyboard_action)
+	virtual bool OnKeyboardAction(u8 dik, EUIMessages keyboard_action)
 	{ return call_member<bool>(this,"OnKeyboard", dik, keyboard_action);}
-	static bool OnKeyboard_static(inherited* ptr, int dik, EUIMessages keyboard_action)
+	static bool OnKeyboard_static(inherited* ptr, u8 dik, EUIMessages keyboard_action)
 	{ 
 		return ptr->self_type::inherited::OnKeyboardAction(dik,keyboard_action );
 	}
