+++ code/engine.vc2008/xrUICore/UIFixedScrollBar.h
@@ -29,6 +29,6 @@ public:
 	virtual bool OnMouseDown(int mouse_btn);
 	virtual	bool OnMouseDownEx();
 	virtual void OnMouseUp(int mouse_btn);
-	virtual bool OnKeyboardHold(int dik);
+	virtual bool OnKeyboardHold(u8 dik);
 
 };
