+++ code/engine.vc2008/xrUICore/UIScrollBar.h
@@ -58,7 +58,7 @@ public:
 	virtual bool 	OnMouseDown			(int mouse_btn);
 	virtual	bool	OnMouseDownEx		();
 	virtual void	OnMouseUp			(int mouse_btn);
-	virtual bool	OnKeyboardHold		(int dik);
+	virtual bool	OnKeyboardHold		(u8 dik);
 
 	virtual void	Draw				();
